+++ test/fixtures/namespace/product/Product.php
@@ -10,6 +10,7 @@ class Product {
     private Name $name;
     private Price $price;
 
+    /** @return Product product */
     public function method1(string $param1) {
     }
 }