+++ test/fixtures/no-namespace/product/Product.php
@@ -3,6 +3,7 @@ class Product {
     private Name $name;
     private Price $price;
 
-    public function method1(string $param1) {
+    /** @return Product product */
+    public function method1(string $param1): Product {
     }
 }