+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+### Added
+
+ * Added dependency arrows. The class has a dependency on the return type of the method with no public arguments.
+
 ## v0.0.1 (2021-11-11)
 
 ### Added